refactor: assets/images 추가
